total admin menu fix (unreliability, yoast seo, ajax)
In 10 late night (or 11 early morning), we got report about admin menu editor. The report was about yoast seo (and bunch of plugins) menu items didn't show up in the builder.

And then we check it backwards through commits & releases...

Previously in 19th of April 2022, we fix un-reliable admin menu issue ( commit #5cc113d ).

The unreliable thing (if we remember it correctly) was because some menu which shouldn't be shown in specific role but they were shown. E.g: subscriber has a lot of menu which shouldn't be shown in editor but they were shown.

Removing the direct recent menu saving (in ajax call) from that commit fix these problems. Also, @MapSteps said, the  extra checking of capability
 (commit #6ec761e) does help.