admin menu refinement (see description)
We used to "constantly" (or kind of) getting issues about admin menu editor (the builder). For instances:
- the order of the admin menu.
- not able to correctly get the admin menu of specific role if there's no user with that role accessing the admin area before.

We need a better, more reliable way to handle the admin menu that will work in "whatever" condition.

What we do here is: temporarily set the `wp_doing_ajax` to false and we do it earlier by setting it directly without being hooked.
With `wp_doing_ajax` is (temporarily & earlier) set to false, we're able to get the admin menu (with correct order) of plugins that do some specific stuff when inside ajax request (such as TablePress & Yoast SEO).

However, we still won't be able to get the menu reliably if:
 * - it is executed earlier than UDB, and they do similar stuff like Yoast SEO
 * - or, the check for DOING_AJAX constant directly instead checking for wp_doing_ajax() (like what Loco Translate does)
But not a porblem, those conditions are already handled by saving the recent menu (see class-admin-menu-module.php).

After this commit, the major issues are fixed, but some minor issues stay (not sure, are they new or already since before):
- under appearance, customizer is double
- under woocommerce, there's product export & product import (only in builder, doesn't affect the output even after saving, which is fine)
- defender pro has a lot of submenus in the builder (only in builder, doesn't affect the output even after saving, which is fine)

Let's handle them in the next commit.