hiding taxonomy submenu items of custom post type in the builder
So, some submenu items may have `&amp;` code instead of `&` sign.
Such as taxonomy submenu items of a custom post type.

However, when rendered into the builder,
the `&amp;` code is converted to `&` sign automatically (browser behavior maybe?).
That url default is rendered in the url field's placeholder & in the `data-default-url` attribute of the submenu item's `li` tag.

That means, even though those taxonomy submenu items of a custom post type
are using `&amp;` code instead of `&` sign, they will be saved as `&` sign in the database.

It will cause a doubled menu items in the array when we try to load it via ajax.
But they will not be doubled in the builder.

This is because the doubled items will be elimited to 1 item in `parse_response_with_custom_menu` execution.
However, the elimination from doubled items into 1 item in that execution resulting wrong result:
The hidden taxonomy submenu item of custom a post type is not hidden in the builder.
That's why we need this extra handling.