makes the codes more maintainable (see description)
Previously, the codes inside preview.js file are too much.
It was quite hard to maintain, to work on.

Now we split them into smaller files.
The build tool is using Parcel.

The codes are also using Typescript (.ts) instead of JS.
This reduces the risk of error in run time (production).
Typescript requires strict-typing that makes the codes much more reliable.